task 6 edit

+++ 0x19-hash_tables/6-hash_table_delete.c
@@ -5,7 +5,7 @@
  */
 void hash_table_delete(hash_table_t *ht)
 {
-	unsigned long int i;
+	unsigned int i;
 	hash_node_t *temp;
 
 	if (ht == NULL || ht->array == NULL)
